+++ resources/views/components/accordion/item.blade.php
@@ -48,6 +48,7 @@ $mergedAttributes = $attributes->merge([
             }
         }
     }"
+    x-cloak
     role="region">
     <button
         type="button"