+++ resources/views/components/alert/index.blade.php
@@ -46,6 +46,7 @@ use SpireUI\Support\ComponentClass;
     @if($closable)
         x-data="{ visible: true }"
         x-show="visible"
+        x-cloak
         x-transition:enter="transition ease-out duration-200"
         x-transition:enter-start="opacity-0 scale-95"
         x-transition:enter-end="opacity-100 scale-100"