+++ resources/views/components/popover/trigger.blade.php
@@ -2,6 +2,7 @@
 
 <div
     x-ref="trigger"
+    @click="toggle()"
     {{ $attributes }}
 >
     {{ $slot }}